Extract handler functions into particular classes
